transform_function/requirements.txt

azure-functions
azure-storage-blob
pyspark>=3.3.0,<3.6.0 